requirements.txt
torch==2.0.1
transformers==4.30.2
Pillow==9.5.0
requests==2.31.0
numpy>=1.21.0